i2c in main geïmplementeerd
