+++ EmbeddedProject/EmbeddedProject/EmbeddedProject.cproj
@@ -131,13 +131,16 @@
     </ToolchainSettings>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="i2c.c">
+      <SubType>compile</SubType>
+    </Compile>
     <Compile Include="i2c.h">
       <SubType>compile</SubType>
     </Compile>
     <Compile Include="main.c">
       <SubType>compile</SubType>
     </Compile>
-    <Compile Include="rp6.c">
+    <Compile Include="main.h">
       <SubType>compile</SubType>
     </Compile>
     <Compile Include="rp6aansluitingen.h">
